// src/app/interfaces/user-profile.ts
export interface UserProfile {
  userId: number;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  location: string;
  locationRadius: number;
  tradeCategory: string;
  tradeSubcategory: string;
  expectedPay: number;
  bankAccountNumber: string;
  bankSortCode: string;
  niNumber: string;
  sex: string;
  rating: number;
}